+++ packages/cheqd-blockchain-modules/src/accumulator/internal.js
@@ -5,62 +5,85 @@ import {
   CheqdAccumulatorParamsRef,
   CheqdAccumulatorWithUpdateInfo,
   CheqdCreateResource,
+  DockAccumulatorWithUpdateInfo,
   CheqdAccumulatorPublicKeyId,
   CheqdAccumulatorParamsId,
-} from '@docknetwork/credential-sdk/types';
-import { TypedUUID, option } from '@docknetwork/credential-sdk/types/generic';
-import { stringToU8a, u8aToString } from '@docknetwork/credential-sdk/utils';
-import { VBWitnessUpdateInfo } from '@docknetwork/credential-sdk/crypto';
+} from "@docknetwork/credential-sdk/types";
+import { TypedUUID, option } from "@docknetwork/credential-sdk/types/generic";
+import { stringToU8a, u8aToString } from "@docknetwork/credential-sdk/utils";
+import { VBWitnessUpdateInfo } from "@docknetwork/credential-sdk/crypto";
 import {
   injectParams,
   injectPublicKeys,
   createInternalCheqdModule,
   SortedResourceVersions,
   validateResource,
-} from '../common';
+} from "../common";
 
-const Type = 'accumulator';
+const Type = "accumulator";
 
 const methods = {
   addAccumulator(accumulatorId, accumulator) {
     const { AccumulatorId, StoredAccumulator } = this.types;
 
     const [did, id] = AccumulatorId.from(accumulatorId).value;
-    const storedAcc = new StoredAccumulator(accumulator).toJSONStringBytes();
+    let storedAcc;
+    let versionId;
+
+    if (accumulator instanceof DockAccumulatorWithUpdateInfo) {
+      storedAcc = new StoredAccumulator(
+        accumulator.accumulator
+      ).toJSONStringBytes();
+      versionId = TypedUUID.fromDockIdent(id, String(accumulator.lastModified));
+    } else {
+      storedAcc = new StoredAccumulator(accumulator).toJSONStringBytes();
+      versionId = TypedUUID.random();
+    }
 
     return new CheqdCreateResource(
       did.value.value,
-      TypedUUID.random(),
-      '1.0',
+      versionId,
+      "1.0",
       [],
       String(id),
       Type,
-      storedAcc,
+      storedAcc
     );
   },
   updateAccumulator(
     accumulatorId,
     accumulator,
-    { additions, removals, witnessUpdateInfo },
+    { additions, removals, witnessUpdateInfo }
   ) {
     const { AccumulatorId, StoredAccumulator } = this.types;
 
     const [did, id] = AccumulatorId.from(accumulatorId).value;
-    const storedAcc = new StoredAccumulator(
-      accumulator,
-      additions,
-      removals,
-      witnessUpdateInfo,
-    ).toJSONStringBytes();
+
+    let storedAcc;
+    let versionId;
+    if (accumulator instanceof DockAccumulatorWithUpdateInfo) {
+      storedAcc = new StoredAccumulator(
+        accumulator.accumulator
+      ).toJSONStringBytes();
+      versionId = TypedUUID.fromDockIdent(id, String(accumulator.lastModified));
+    } else {
+      storedAcc = new StoredAccumulator(
+        accumulator,
+        additions,
+        removals,
+        witnessUpdateInfo
+      ).toJSONStringBytes();
+      versionId = TypedUUID.random();
+    }
 
     return new CheqdCreateResource(
       did.value.value,
-      TypedUUID.random(),
-      '1.0',
+      versionId,
+      "1.0",
       [],
       String(id),
       Type,
-      storedAcc,
+      storedAcc
     );
   },
   removeAccumulator(accumulatorId) {
@@ -71,24 +94,24 @@ const methods = {
     return new CheqdCreateResource(
       did.value.value,
       TypedUUID.random(),
-      '1.0',
+      "1.0",
       [],
       String(id),
       Type,
-      stringToU8a('null'),
+      stringToU8a("null")
     );
   },
 };
 
 export default class CheqdInternalAccumulatorModule extends injectParams(
-  injectPublicKeys(createInternalCheqdModule(methods)),
+  injectPublicKeys(createInternalCheqdModule(methods))
 ) {
   static get MsgNames() {
     return {
       ...super.MsgNames,
-      addAccumulator: 'MsgCreateResource',
-      updateAccumulator: 'MsgCreateResource',
-      removeAccumulator: 'MsgCreateResource',
+      addAccumulator: "MsgCreateResource",
+      updateAccumulator: "MsgCreateResource",
+      removeAccumulator: "MsgCreateResource",
     };
   }
 
@@ -98,24 +121,72 @@ export default class CheqdInternalAccumulatorModule extends injectParams(
 
   static PublicKeyId = CheqdAccumulatorPublicKeyId;
 
-  static PublicKeyName = 'AccumulatorPublicKey';
+  static PublicKeyName = "AccumulatorPublicKey";
 
-  static PublicKeyType = 'accumulator-public-key';
+  static PublicKeyType = "accumulator-public-key";
 
   static ParamsId = CheqdAccumulatorParamsId;
 
   static Params = AccumulatorParams;
 
-  static ParamsName = 'AccumulatorParams';
+  static ParamsName = "AccumulatorParams";
 
-  static ParamsType = 'accumulator-params';
+  static ParamsType = "accumulator-params";
 
   static ParamsRef = CheqdAccumulatorParamsRef;
 
   createAccumulatorMetadataFilter(name) {
     const strName = String(name);
 
-    return (meta) => meta.resourceType === 'accumulator' && meta.name === strName;
+    return (meta) =>
+      meta.resourceType === "accumulator" && meta.name === strName;
+  }
+
+  async accumulatorHistory(accumulatorId) {
+    const [did, name] = CheqdAccumulatorId.from(accumulatorId).value;
+
+    const ids = new SortedResourceVersions(
+      await this.resourcesMetadataBy(
+        did,
+        this.createAccumulatorMetadataFilter(name)
+      )
+    ).ids();
+
+    const accumulators = [
+      ...new SortedResourceVersions(await this.resources(did, ids)),
+    ].map((acc) =>
+      CheqdStoredAccumulator.from(
+        JSON.parse(u8aToString(validateResource(acc, String(name), Type)))
+      )
+    );
+
+    return {
+      created: new CheqdAccumulatorWithUpdateInfo(
+        ids[0],
+        ids[0],
+        accumulators[0].accumulator
+      ),
+      updates: accumulators.slice(1).map((acc) => ({
+        newAccumulated: acc.accumulated,
+        additions: acc.additions
+          ? acc.additions.map((addition) => addition.bytes)
+          : [],
+        removals: acc.removals
+          ? acc.removals.map((removal) => removal.bytes)
+          : [],
+        witnessUpdateInfo: acc.witnessUpdateInfo ?? new Uint8Array(),
+      })),
+    };
+  }
+
+  async lastAccumulatorId(accumulatorId) {
+    const [did, name] = CheqdAccumulatorId.from(accumulatorId).value;
+    const lastMeta = await this.latestResourceMetadataBy(
+      did,
+      this.createAccumulatorMetadataFilter(name)
+    );
+
+    return lastMeta?.id;
   }
 
   async accumulator(accumulatorId) {
@@ -123,8 +194,8 @@ export default class CheqdInternalAccumulatorModule extends injectParams(
     const ids = new SortedResourceVersions(
       await this.resourcesMetadataBy(
         did,
-        this.createAccumulatorMetadataFilter(name),
-      ),
+        this.createAccumulatorMetadataFilter(name)
+      )
     ).ids();
 
     if (!ids.length) {
@@ -137,10 +208,10 @@ export default class CheqdInternalAccumulatorModule extends injectParams(
           validateResource(
             await this.resource(did, ids[ids.length - 1]),
             String(name),
-            Type,
-          ),
-        ),
-      ),
+            Type
+          )
+        )
+      )
     );
 
     if (acc == null) {
@@ -150,14 +221,14 @@ export default class CheqdInternalAccumulatorModule extends injectParams(
     return new CheqdAccumulatorWithUpdateInfo(
       ids[0],
       ids[ids.length - 1],
-      acc.accumulator,
+      acc.accumulator
     );
   }
 
   async lastParamsId(did) {
     const res = await this.latestResourceMetadataBy(
       did,
-      this.filterParamsMetadata,
+      this.filterParamsMetadata
     );
 
     return res?.id;
@@ -166,7 +237,7 @@ export default class CheqdInternalAccumulatorModule extends injectParams(
   async lastPublicKeyId(did) {
     const res = await this.latestResourceMetadataBy(
       did,
-      this.filterPublicKeyMetadata,
+      this.filterPublicKeyMetadata
     );
 
     return res?.id;
@@ -187,7 +258,7 @@ export default class CheqdInternalAccumulatorModule extends injectParams(
     member,
     witness,
     start,
-    end,
+    end
   ) {
     const [did, name] = CheqdAccumulatorId.from(accumulatorId).value;
     const startUUID = String(TypedUUID.from(start));
@@ -196,8 +267,8 @@ export default class CheqdInternalAccumulatorModule extends injectParams(
     const sortedIDs = new SortedResourceVersions(
       await this.resourcesMetadataBy(
         did,
-        this.createAccumulatorMetadataFilter(name),
-      ),
+        this.createAccumulatorMetadataFilter(name)
+      )
     ).ids();
 
     const startIdx = sortedIDs.findIndex((id) => id === startUUID);
@@ -207,7 +278,7 @@ export default class CheqdInternalAccumulatorModule extends injectParams(
 
       if (endIdx === -1) {
         throw new Error(
-          `Accumulator \`${accumulatorId}\` with version \`${end}\` doesn't exist`,
+          `Accumulator \`${accumulatorId}\` with version \`${end}\` doesn't exist`
         );
       }
     } else {
@@ -216,19 +287,20 @@ export default class CheqdInternalAccumulatorModule extends injectParams(
 
     const accumulators = await this.resources(
       did,
-      sortedIDs.slice(startIdx, endIdx + 1),
+      sortedIDs.slice(startIdx, endIdx + 1)
     );
 
     for (const accumulator of new SortedResourceVersions(accumulators)) {
-      const { additions, removals, witnessUpdateInfo } = CheqdStoredAccumulator.from(
-        validateResource(accumulator, String(name), Type),
-      );
+      const { additions, removals, witnessUpdateInfo } =
+        CheqdStoredAccumulator.from(
+          validateResource(accumulator, String(name), Type)
+        );
 
       witness.updateUsingPublicInfoPostBatchUpdate(
         member,
         additions ? [...additions].map((addition) => addition.bytes) : [],
         removals ? [...removals].map((removal) => removal.bytes) : [],
-        new VBWitnessUpdateInfo(witnessUpdateInfo?.bytes || new Uint8Array()),
+        new VBWitnessUpdateInfo(witnessUpdateInfo?.bytes || new Uint8Array())
       );
     }
   }