+++ scripts/migration/src/dock-id-to-cheqd.js
@@ -4,8 +4,10 @@ import {
   DockAccumulatorId,
   DockBlobId,
   DockStatusListCredentialId,
+  DockDid,
 } from "@docknetwork/credential-sdk/types";
 import pLimit from "p-limit";
+import { DockAccumulatorModule } from "@docknetwork/dock-blockchain-modules";
 
 const mergeAwait = async (objectPromises) => {
   const acc = {};